+++ App.py
@@ -1,35 +1,9 @@
 import streamlit as st
 import pandas as pd
 import matplotlib.pyplot as plt
-import psycopg
-import os
-from dotenv import load_dotenv
+from lunch_menu.db import get_connection, db_name
 
 members = { "TOM" : 1, "cho" : 2, "hyun" : 3, "JERRY" : 4, "SEO" : 5, "jiwon" : 6, "jacob" : 7, "heejin" : 8, "lucas" : 9, "nuni" : 10 }
-
-# https://docs.streamlit.io/develop/concepts/connections/secrets-management
-# DB_CONFIG = {
-   # "dbname": "sunsindb",
-   # "user": "sunsin",
-   # "password": "mysecretpassword",
-   # "host": "localhost",
-   # "port": "5432",
-#}
-# 위 정보를 안보이게 하는 방법은 아래 참고
-
-load_dotenv()
-db_name = os.getenv("DB_NAME")
-DB_CONFIG = {
-    #"user": st.secrets["db_username"]    
-    "dbname": os.getenv("DB_NAME"),
-    "user": os.getenv("DB_USERNAME"),
-    "password": os.getenv("DB_PASSWORD"),
-    "host": os.getenv("DB_HOST"),
-    "port": os.getenv("DB_PORT"),
-}
-
-def get_connection():
-    return psycopg.connect(**DB_CONFIG)
 
 def insert_menu(menu_name, member_id, dt):
     try:
